+++ tests/run-tests.js
@@ -7,6 +7,7 @@ const sourceFiles = [
   'src/core/html-utils.js',
   'src/core/dsl-utils.js',
   'src/core/regex-parts.js',
+  'src/core/id-normalizer.js',
   'src/core/line-resolver.js',
   'src/core/formatter-interface.js',
   'src/core/dsl-updater.js',