+++ tests/sequence-overlay.test.js
@@ -20,6 +20,7 @@ delete require.cache[require.resolve('../src/core/selection-router.js')];
 require('../src/core/html-utils.js');
 require('../src/core/dsl-utils.js');
 require('../src/core/regex-parts.js');
+require('../src/core/id-normalizer.js');
 require('../src/core/dsl-updater.js');
 require('../src/core/text-updater.js');
 require('../src/core/parser-utils.js');